Moved simple executor to Ktor plugin
