+++ ktor/server/src/main/kotlin/org/merideum/ktor/server/executor/SimpleDependency.kt
@@ -1,4 +1,4 @@
-package org.merideum.kotlin.merit.execution
+package org.merideum.ktor.server.executor
 
 import org.merideum.kotlin.merit.interpreter.Dependency
